refactor: move duplicate block to same level
