+++ apps/journeys-admin/src/components/Editor/EditToolbar/DuplicateBlock/DuplicateBlock.spec.tsx
@@ -10,12 +10,12 @@ import {
   GetJourney_journey as Journey,
   GetJourney_journey_blocks_StepBlock as StepBlock,
   GetJourney_journey_blocks_TypographyBlock as TypographyBlock
-} from '../../../__generated__/GetJourney'
+} from '../../../../../__generated__/GetJourney'
 import {
   TypographyAlign,
   TypographyColor,
   TypographyVariant
-} from '../../../__generated__/globalTypes'
+} from '../../../../../__generated__/globalTypes'
 
 import { BLOCK_DUPLICATE, DuplicateBlock } from './DuplicateBlock'
