+++ apps/journeys-admin/src/components/Editor/EditToolbar/EditToolbar.tsx
@@ -12,7 +12,7 @@ import {
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
 import EyeOpenIcon from '@core/shared/ui/icons/EyeOpen'
 
-import { DuplicateBlock } from '../../DuplicateBlock'
+import { DuplicateBlock } from './DuplicateBlock'
 
 import { Analytics } from './Analytics'
 import { DeleteBlock } from './DeleteBlock'
