+++ apps/journeys-admin/src/components/Editor/EditToolbar/Menu/Menu.tsx
@@ -19,7 +19,7 @@ import SettingsIcon from '@core/shared/ui/icons/Settings'
 import { GetRole } from '../../../../../__generated__/GetRole'
 import { Role } from '../../../../../__generated__/globalTypes'
 import { setBeaconPageViewed } from '../../../../libs/setBeaconPageViewed'
-import { DuplicateBlock } from '../../../DuplicateBlock'
+import { DuplicateBlock } from '../DuplicateBlock'
 import { MenuItem } from '../../../MenuItem'
 import { Analytics } from '../Analytics'
 import { DeleteBlock } from '../DeleteBlock'
